+++ changelog.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Get all method
+- Delete all method
 - Default lambda input value for update method
 
 ## [0.3.0] - 2021-06-04